requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=1.8.0,<2.0.0
python-multipart==0.0.6
rich==13.7.0
typer==0.9.0
requests==2.31.0
websockets==12.0
openai==1.3.0 